+++ core/data/src/main/kotlin/dev/a10101100/snipcraft/core/data/DataModule.kt
@@ -1,0 +1,20 @@
+package dev.a10101100.snipcraft.core.data
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class DataModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindSnippetRepository(impl: SnippetRepositoryImpl): SnippetRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindFolderRepository(impl: FolderRepositoryImpl): FolderRepository
+}